Rename the module from openmaple to maplesoft.maple.
