+++ pyproject.toml
@@ -24,10 +24,10 @@ classifiers = [
 ]
 license = { file = "LICENSE.txt" }
 
+dependencies = [
+    "numpy >= 1.23",
+]
 [project.urls]
 Homepage = "https://github.com/Maplesoft/openmaple"
 Issues = "https://github.com/Maplesoft/openmaple/issues"
 
-dependencies = [
-    "numpy >= 1.23",
-]
